Refactor ConnectionElement to have a front and backend
