+++ mfp/gui/modes/global_mode.py
@@ -122,7 +122,7 @@ class GlobalMode (InputMode):
     async def transient_msg(self):
         from ..message_element import TransientMessageElement
         if self.window.selected:
-            return await self.window.add_element(TransientMessageElement.get_factory())
+            return await self.window.add_element(TransientMessageElement.build)
         return False
 
     async def hover(self, details):
@@ -212,18 +212,21 @@ class GlobalMode (InputMode):
         if select_mode is None:
             if self.manager.pointer_obj is not None:
                 if self.manager.pointer_obj not in self.window.selected:
+                    log.debug(f"[selbox] selecting pointer_obj {self.manager.pointer_obj}")
                     await self.window.unselect_all()
                     await self.window.select(self.manager.pointer_obj)
                     raise InputManager.InputNeedsRequeue()
+                else:
+                    log.debug(f"[selbox] was None, pointer_obj={self.manager.pointer_obj}, selected={self.window.selected}")
 
                 if self.allow_selection_drag:
                     self.selection_drag_started = True
             else:
+                log.debug(f"[selbox] pointer_obj={self.manager.pointer_obj}, selected={self.window.selected}")
                 await self.window.unselect_all()
                 self.selbox_started = True
         elif select_mode is True:
-            if (self.manager.pointer_obj
-                    and self.manager.pointer_obj not in self.window.selected):
+            if (self.manager.pointer_obj and self.manager.pointer_obj not in self.window.selected):
                 await self.window.select(self.manager.pointer_obj)
             self.selbox_started = True
         else:
