+++ mfp/gui/text_element.py
@@ -45,7 +45,7 @@ class TextElement (ClutterBaseElementBackend):
         self.texture.connect("draw", self.draw_cb)
         self.group.set_content(self.texture)
 
-        self.label = TextWidget.get_factory()(self)
+        self.label = TextWidget.build(self)
         self.label.set_color(self.get_color('text-color'))
         self.label.set_font_name(self.get_fontspec())
         self.label.set_position(3, 3)
